chore(backend): fix orphans rules problems
