+++ apps/backend/src/main.rs
@@ -29,7 +29,9 @@ use itertools::Itertools;
 use logs_wheel::LogFileInitializer;
 use migrations::Migrator;
 use models::{prelude::Exercise, CompleteExport};
-use sea_orm::{ConnectOptions, ConnectionTrait, Database, DatabaseConnection, PaginatorTrait};
+use sea_orm::{
+    ConnectOptions, ConnectionTrait, Database, DatabaseConnection, EntityTrait, PaginatorTrait,
+};
 use sea_orm_migration::MigratorTrait;
 use tokio::{
     join,
